+++ SC_Synthetic_UH_Method.py
@@ -1,8 +1,72 @@
 import requests
 import ast
 import numpy as np
-from Rainfall_Data_Curves import rainfall_data_curves
 import math
+from Tc_Calculator import lagTimeMethodTimeOfConcentration, travelTimeMethodTimeOfConcentration
+from Rainfall_Data_Curves import rainfall_data_curves
+
+
+# Combines rainfallDistributionCurve, PRFData, weightedCurveNumber, and travelTimeMethodTimeOfConcentration or lagTimeMethodTimeOfConcentration (depending on TcMethod) into single function.
+def calculateMissingParametersSCSUH(lat, lon, AEP, curveNumberMethod, TcMethod, length=None, slope=None, dataSheetFlow=None, dataExcessSheetFlow=None, dataShallowConcentratedFlow=None, dataChannelizedFlowOpenChannel=None, dataChannelizedFlowStormSewer=None, dataChannelizedFlowStormSewerOrOpenChannelUserInputVelocity=None):
+    # AEP: 10 - 10 year return period, 4 - 25 year return period, 2 - 50 year return period, 1 - 100 year return period
+    # curveNumberMethod: "runoff" or "area"
+    # TcMethod: 'lagtime' or 'traveltime'
+    # length: flow path length in feet
+    # slope: flow path slope in %
+    # dataSheetFlow: data corresponding to Sheet Flow section for Travel Time Method
+    # dataExcessSheetFlow: data corresponding to Excess Sheet Flow section for Travel Time Method 
+    # dataShallowConcentratedFlow: data corresponding to Shallow Concentrated Flow section for Travel Time Method
+    # dataChannelizedFlowOpenChannel: data corresponding to Channelized Flow - Open Channel section for Travel Time Method
+    # dataChannelizedFlowStormSewer: data corresponding to Channelized Flow - Storm Sewer section for Travel Time Method
+    # dataChannelizedFlowStormSewerOrOpenChannelUserInputVelocity: data corresponding to Channelized Flow (Storm Sewer and/or Open Channel) - User Input Velocity section for Travel Time Method
+
+    # Get Rainfall Distribution Curve letter
+    rainfall_distribution_curve = rainfallDistributionCurve(lat, lon) # Get from rainfallDistributionCurve function
+    rainfall_distribution_curve_letter = rainfall_distribution_curve[0]
+    
+    # Get rainfallData
+    rainfall_data = rainfallData(lat, lon)
+
+    # Get Tc
+    if TcMethod.lower() == "traveltime":
+        if all([dataSheetFlow, dataExcessSheetFlow, dataShallowConcentratedFlow, dataChannelizedFlowStormSewer, dataChannelizedFlowStormSewerOrOpenChannelUserInputVelocity]):
+            P2_24_2 = rainfall_data[25]
+            Tc = travelTimeMethodTimeOfConcentration(dataSheetFlow, dataExcessSheetFlow, P2_24_2,
+                                        dataShallowConcentratedFlow,
+                                        dataChannelizedFlowOpenChannel,
+                                        dataChannelizedFlowStormSewer,
+                                        dataChannelizedFlowStormSewerOrOpenChannelUserInputVelocity)
+            Tc = {'value': Tc[0],'warningMessage': Tc[1] }
+        else:
+            raise Exception("Not all parameters for traveltime were entered.")
+    elif TcMethod.lower() == "lagtime":
+        if all([length, slope]):
+            rainfall_distribution_curve_number = rainfall_distribution_curve[1]
+            Tc = lagTimeMethodTimeOfConcentration(length, slope, rainfall_distribution_curve_number)
+            Tc = {'value': Tc,'warningMessage': None }
+
+        else:
+            raise Exception("Not all parameters for lagtime were entered.")
+    else:
+        raise Exception("Time of concentration method not valid.")
+    
+    # Get PRF
+    PRF = PRFData (lat, lon)
+    
+    # Get Curve Number, S, Ia
+    if curveNumberMethod.lower() == "runoff" or curveNumberMethod.lower() == "area":
+        if AEP == 1 or AEP == 2 or AEP == 4 or AEP == 10:
+            if AEP == 1:  P24hr = rainfall_data[23]
+            if AEP == 2:  P24hr = rainfall_data[17]
+            if AEP == 4:  P24hr = rainfall_data[11]
+            if AEP == 10:  P24hr = rainfall_data[5]
+            CN, S, Ia = weightedCurveNumber(lat, lon, P24hr, curveNumberMethod)
+        else:
+            raise Exception("AEP not valid.")
+    else:
+        raise Exception("Curve number method not valid.")
+
+    return rainfall_distribution_curve_letter, Tc, PRF, CN, S, Ia
 
 # Extracts data from curve number GIS layer, then computes Runoff Weighted CN or Area Weighted CN
 # Corresponds to "Data for CN Determination" sheet in spreadsheet